requirements.txt
Flask
langchain-openai
langchain_experimental
pinecone-client
